quantumsafenet-backend/requirements.txt
uvicorn
fastapi
pydantic
pydantic[email]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
pymysql
python-multipart
requests
python-socketio
fastapi-socketio